# Fix _split_polygon_with_line performance for big polygons with many holes